+++ src/main/kotlin/com/example/team/haribo/goms/gomsserverv3/global/security/JwtAuthenticationFilter.kt
@@ -1,19 +1,17 @@
-package com.example.team.haribo.goms.GomsServerV3.global.security
+package com.example.team.haribo.goms.gomsserverv3.global.security
 
-import com.example.team.haribo.goms.GomsServerV3.global.exception.ErrorCode
-import com.example.team.haribo.goms.GomsServerV3.global.exception.ErrorResponse
-import com.example.team.haribo.goms.GomsServerV3.global.jwt.JwtProvider
-import com.fasterxml.jackson.databind.ObjectMapper
+import com.example.team.haribo.goms.gomsserverv3.global.exception.ErrorCode
+import com.example.team.haribo.goms.gomsserverv3.global.jwt.JwtProvider
 import io.jsonwebtoken.ExpiredJwtException
 import io.jsonwebtoken.JwtException
 import jakarta.servlet.FilterChain
 import jakarta.servlet.http.HttpServletRequest
 import jakarta.servlet.http.HttpServletResponse
-import org.springframework.http.MediaType
 import org.springframework.security.authentication.UsernamePasswordAuthenticationToken
 import org.springframework.security.core.authority.SimpleGrantedAuthority
 import org.springframework.security.core.context.SecurityContextHolder
 import org.springframework.web.filter.OncePerRequestFilter
+import tools.jackson.databind.ObjectMapper
 
 class JwtAuthenticationFilter(
     private val jwtProvider: JwtProvider,
@@ -29,10 +27,16 @@ class JwtAuthenticationFilter(
 
         if (token != null) {
             try {
-                jwtProvider.validate(token)
+                val claims = jwtProvider.parseClaims(token)
 
-                val memberId = jwtProvider.getMemberId(token)
-                val role = jwtProvider.getRole(token) ?: "ROLE_STUDENT"
+                val type = claims["type"]?.toString()
+                if (type != "ACCESS") {
+                    SecurityErrorResponseWriter.write(response, objectMapper, ErrorCode.INVALID_TOKEN)
+                    return
+                }
+
+                val memberId = claims.subject.toLong()
+                val role = claims["role"]?.toString() ?: "ROLE_STUDENT"
 
                 val authentication = UsernamePasswordAuthenticationToken(
                     memberId,
@@ -42,13 +46,13 @@ class JwtAuthenticationFilter(
 
                 SecurityContextHolder.getContext().authentication = authentication
             } catch (e: ExpiredJwtException) {
-                writeError(response, ErrorCode.EXPIRED_TOKEN)
+                SecurityErrorResponseWriter.write(response, objectMapper, ErrorCode.EXPIRED_TOKEN)
                 return
             } catch (e: JwtException) {
-                writeError(response, ErrorCode.INVALID_TOKEN)
+                SecurityErrorResponseWriter.write(response, objectMapper, ErrorCode.INVALID_TOKEN)
                 return
             } catch (e: IllegalArgumentException) {
-                writeError(response, ErrorCode.INVALID_TOKEN)
+                SecurityErrorResponseWriter.write(response, objectMapper, ErrorCode.INVALID_TOKEN)
                 return
             }
         }
@@ -60,19 +64,5 @@ class JwtAuthenticationFilter(
         val header = request.getHeader("Authorization") ?: return null
         if (!header.startsWith("Bearer ")) return null
         return header.substring(7).trim().ifEmpty { null }
-    }
-
-    private fun writeError(response: HttpServletResponse, errorCode: ErrorCode) {
-        response.status = errorCode.status
-        response.contentType = MediaType.APPLICATION_JSON_VALUE
-        response.characterEncoding = "UTF-8"
-        response.writer.write(
-            objectMapper.writeValueAsString(
-                ErrorResponse(
-                    status = errorCode.status,
-                    message = errorCode.message
-                )
-            )
-        )
     }
 }
